Adding pending tasks metadata
Vyper is still under heavy development and as such there are a few areas which I would like to keep an eye on (rather than writing about redundant or deprecated information). 
I have added revision number, revision date, revision remark and revision status to keep track of the research and development of each Vyper subsection.
These will all be updated with the latest available information in time for ethereumbook's publication.