Adjustments to wording of programming paradigm
Rewording the paragraph which suggests that we may be in the midst of a new software programming paradigm known as the immutability paradigm. 
Errors in blockchain code can and have resulted in hard forks. This is undesirable.
On the flip side, in most (if not all) NON-blockchain settings system attributes, properties and variables which hold value are ultimately alterable (given the right administrative privileges and permissions).
It is suggested that the blockchain revolution could be opening new opportunities for further research and development into an immutability paradigm.