What is ethereum comments (#2)
* comments to what is ethereum chapter

Some suggestions on how to make this intro chapter more complete.

I used the 4/ to insert the comments, I hope that this is the format you would expect, if not please advise.

As a side comment for the rest of the book, I read few chapters and I had the feeling that sometimes the authors seem to be shy talking about bitcoin directly (instead referring to hypothetical blockchains). Since everyone knows Andreas’ expertise and passion about bitcoin, imho nobody would think that it’s not politically correct to use the word bitcoin as this is also the obvious benchmark. It would actually perfectly fit with Andreas’ intellectual honesty and possibly help the community to mature.

 Possibly a dedicated chapter about the lion vs shark topic could fit in the book, also from a programmer’s perspective.