Responding to comment about modifiers
Taking some good advice which was left in a comment. I have responded to the commend inline and then also rewritten the paragraph in question. I hope this helps the reader in grasping a better understanding of why Vyper does not support modifiers.

Also adding a GitHub URL to the explicit logic of decorators (namely the function_signature.py file). This is so that readers can click through and view how Vyper applies decorators (code also shows exceptions and error messages which could be very useful for new developers).